% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_catalog}
\alias{get_catalog}
\title{Fetch the dataset catalog}
\usage{
get_catalog()
}
\value{
A named list mapping dataset names to metadata URLs.
}
\description{
Fetch the dataset catalog
}
\keyword{internal}
