% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_connection}
\alias{get_connection}
\title{Get a DuckDB connection configured for R2}
\usage{
get_connection()
}
\value{
A DuckDB connection object.
}
\description{
Get a DuckDB connection configured for R2
}
\keyword{internal}
