% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{has_connection}
\alias{has_connection}
\title{Check if a connection can be established}
\usage{
has_connection()
}
\value{
\code{TRUE} if a connection is available, \code{FALSE} otherwise.
}
\description{
Check if a connection can be established
}
\keyword{internal}
