% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_datasets.R
\name{list_datasets}
\alias{list_datasets}
\title{List available datasets in the Bedrock Bio library}
\usage{
list_datasets()
}
\value{
A character vector of dataset names
}
\description{
List available datasets in the Bedrock Bio library
}
\examples{
\dontrun{
library(bedrockbio)
list_datasets()
}

}
