% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_dataset.R
\name{load_dataset}
\alias{load_dataset}
\title{Lazily query a dataset}
\usage{
load_dataset(name)
}
\arguments{
\item{name}{Dataset identifier (e.g., "ukb_ppp.pqtls")}
}
\value{
A lazy \code{tbl} backed by DuckDB, compatible with dplyr verbs.
}
\description{
Lazily query a dataset
}
\examples{
\dontshow{if (bedrockbio:::has_connection()) withAutoprint(\{ # examplesIf}
library(bedrockbio)
library(dplyr)

df <- load_dataset("ukb_ppp.pqtls") |>
  filter(
    ancestry == "EUR",
    protein_id == "A0FGR8"
  ) |>
  select(
    chromosome,
    position,
    effect_allele,
    other_allele,
    beta,
    neg_log_10_p_value
  ) |>
  collect()
\dontshow{\}) # examplesIf}
}
