% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_dataset.R
\name{load_dataset}
\alias{load_dataset}
\title{Lazily read a dataset from the Bedrock Bio library}
\usage{
load_dataset(name)
}
\arguments{
\item{name}{Dataset name (e.g., "ukb_ppp/pqtls")}
}
\value{
A lazy tibble
}
\description{
Lazily read a dataset from the Bedrock Bio library
}
\examples{
\dontrun{
library(bedrockbio)
library(dplyr)

df <- load_dataset("ukb_ppp/pqtls") |>
  filter(
    ancestry == "EUR",
    protein == "A0FGR8"
  ) |>
  select(
    chromosome,
    position,
    effect_allele,
    other_allele,
    beta,
    neg_log_10_p_value
  ) |>
  collect()
}

}
